Unified SceneChangeController and ScenePlayerController
