+++ HeartsOfInk/Assets/Scripts/Controller/Generic/SceneChangeController.cs
@@ -4,13 +4,47 @@ using UnityEngine.SceneManagement;
 
 public class SceneChangeController : MonoBehaviour
 {
+    private string previousSceneKey = "PreviousScene";
+
+    /// <summary>
+    /// A button to go to multiplayer lobby scene.
+    /// </summary>
     public Transform multiplayerButton;
+
+    /// <summary>
+    /// A button to go to credits scene.
+    /// </summary>
     public Transform creditsButton;
+
+    /// <summary>
+    /// A button to return to main menu scene.
+    /// </summary>
     public Transform mainMenuButton;
+
+    /// <summary>
+    /// A button to return to main menu scene.
+    /// </summary>
     public Transform exitToMenuButton;
+
+    /// <summary>
+    /// A button to go to start scene.
+    /// </summary>
     public Transform startGame;
+
+    /// <summary>
+    /// A button to go to options scene.
+    /// </summary>
     public Transform optionsButton;
+
+    /// <summary>
+    /// A button to go to map editor scene.
+    /// </summary>
     public Transform mapEditorButton;
+
+    /// <summary>
+    /// A button to go to login scene.
+    /// </summary>
+    public Transform loginButton;
 
     public enum Scenes
     {
@@ -23,16 +57,18 @@ public class SceneChangeController : MonoBehaviour
         MapEditor = 6,
         AcceptPolicy = 7,
         Options = 8,
-        GameUpdaterScene = 9
+        GameUpdaterScene = 9,
+        Login = 10
     }
 
-    public void DirectChangeScene(Scenes scene)
-    {
-        SceneManager.LoadScene(Convert.ToInt32(scene));
-    }
-
+    /// <summary>
+    /// Used for change scene when specific button is pressed and no specific logic is needed.
+    /// </summary>
+    /// <param name="orderButton"></param>
     public void ChangeScene(Transform orderButton)
     {
+        PlayerPrefs.SetString(previousSceneKey, SceneManager.GetActiveScene().name);
+
         if (AreEquals(orderButton, creditsButton))
         {
             SceneManager.LoadScene(Convert.ToInt32(Scenes.Credits));
@@ -58,15 +94,33 @@ public class SceneChangeController : MonoBehaviour
         {
             SceneManager.LoadScene(Convert.ToInt32(Scenes.Options));
         }
+        else if (AreEquals(orderButton, loginButton))
+        {
+            SceneManager.LoadScene(Convert.ToInt32(Scenes.Login));
+        }
         else
         {
             Debug.LogWarning("Order button unknown: " + orderButton.name);
         }
     }
 
+    /// <summary>
+    /// Used when specific logic is needed for change scene or when change scene isn't started from an user interaction.
+    /// </summary>
+    /// <param name="newScene"></param>
     public void ChangeScene(Scenes newScene)
     {
+        PlayerPrefs.SetString(previousSceneKey, SceneManager.GetActiveScene().name);
         SceneManager.LoadScene(Convert.ToInt32(newScene));
+    }
+
+    public void ChangeToPreviousScene()
+    {
+        string previousScene = PlayerPrefs.GetString(previousSceneKey, string.Empty);
+        if (!string.IsNullOrEmpty(previousScene))
+        {
+            SceneManager.LoadScene(previousScene);
+        }
     }
 
     public void ExitGame()
