--- HeartsOfInk/Assets/Scripts/Controller/Generic/ScenePlayerChange.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7518a8d065c99ef469b0d2516dfe7b0f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
